Improve passport.js JWT implementation
